Add hook for fetching user
